% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{Plot a Riksbank series}
\usage{
plot_series(ts, from = "1900-01-01", to = Sys.Date())
}
\arguments{
\item{ts}{A single series identifier, such as "SECBREPOEFF".
Leading and trailing whitespace is removed.}

\item{from}{The start date, supplied as a Date object or a character
string in YYYY-MM-DD format. Defaults to "1900-01-01".}

\item{to}{The end date, supplied as a Date object or a character
string in YYYY-MM-DD format. Defaults to the current date.
Must be later than or equal to from.}
}
\value{
A ggplot object.
}
\description{
Plot a Riksbank series
}
